+++ apps/api/urls/branch_urls.py
@@ -6,4 +6,5 @@ app_name = 'Branchs endpoints'
 urlpatterns = [
     path('read/list/<int:pk>', views.read_branch_list_view, name='read_list'),
     path('read/<int:pk>', views.read_branch_view, name='read'),
+    path('create/', views.create_branch_view, name='create')
 ]