+++ apps/api/views/user_views.py
@@ -121,7 +121,7 @@ def register_user_view(request):
     try:
         user = create_user_by_data_service(data)
         serializer = UserSerializerWithToken(user, many=False)
-        return Response(serializer.data)
+        return Response(serializer.data, status=status.HTTP_201_CREATED)
 
     except KeyError as e:
         message = f'Ошибка при обработке запроса. Отсутствует поле {e}'