+++ CHANGELOG.md
@@ -6,11 +6,12 @@
 
 ### Добавлено
 
--
+- New endpoint (branch create) [@fringer2423](https://github.com/fringer2423)
 
 ### Изменено
 
 - Edited Branch model fields requirements [@fringer2423](https://github.com/fringer2423)
+- Edited all views (replace 200 -> 201 status)
 
 ## [0.0.21] - 2022-11-03